MAF demo
https://msdn.microsoft.com/en-gb/library/bb384240(v=vs.100).aspx
---------
For the .NET Framework to discover pipeline segments and activate
add-ins, the pipeline segments must be put in a specified directory. The
specified directory names are required but are not case-sensitive. The
only names that are not specified are the name of the pipeline root
directory (which you provide to the discovery methods), and the names of
the subdirectories that contain the add-ins. All the specified segment
names must be subdirectories on the same level under the pipeline root.